Propagate final edition victory state
